+++ src/sub-clients/subjects/SubjectsConverter.ts
@@ -1,14 +1,21 @@
+import { Link } from '../common/model/LinkEntity';
 import { Subject } from './model/Subject';
+import { SubjectLinks } from './model/SubjectLinks';
 
 export class SubjectsConverter {
   public static convert(response: any): Subject {
+    const links: SubjectLinks = {};
+    if (response._links) {
+      if (response._links.update && response._links.update.href) {
+        links.update = new Link({ href: response._links.update.href });
+      }
+    }
+
     return {
       id: response.id,
       name: response.name,
-      updateLink:
-        response._links && response._links.update && response._links.update.href
-          ? response._links.update.href
-          : undefined,
+      lessonPlan: response.lessonPlan,
+      links,
     };
   }
 }