+++ src/sub-clients/subjects/client/FakeSubjectsClient.ts
@@ -1,5 +1,5 @@
 import { Clearable } from '../../common/utils/Clearable';
-import { Subject } from '../model/Subject';
+import { Subject, UpdateSubjectRequest } from '../model/Subject';
 import { SubjectsClient } from './SubjectsClient';
 
 export class FakeSubjectsClient implements SubjectsClient, Clearable {
@@ -12,13 +12,16 @@ export class FakeSubjectsClient implements SubjectsClient, Clearable {
     return Promise.resolve(this.subjects);
   }
 
-  public async update(currentSubject: Subject, newName: string): Promise<void> {
-    if (!currentSubject.updateLink) {
+  public async update(
+    currentSubject: Subject,
+    updateRequest: UpdateSubjectRequest,
+  ): Promise<void> {
+    if (!currentSubject.links.update) {
       throw new Error('Update link not available');
     }
 
     const subject = this.subjects.find(it => it.id === currentSubject.id);
-    subject.name = newName;
+    subject.name = updateRequest.name;
   }
 
   public clear() {