+++ src/sub-clients/subjects/model/Subject.ts
@@ -1,5 +1,12 @@
+import { SubjectLinks } from './SubjectLinks';
+
 export interface Subject {
   id: string;
   name: string;
-  updateLink?: string;
+  lessonPlan: boolean;
+  links?: SubjectLinks;
+}
+
+export interface UpdateSubjectRequest {
+  name: string;
 }