+++ src/sub-clients/subjects/model/SubjectLinks.ts
@@ -1,0 +1,5 @@
+import { Link } from '../../common/model/LinkEntity';
+
+export interface SubjectLinks {
+  update?: Link;
+}