+++ src/test-support/index.ts
@@ -1,4 +1,13 @@
+import { ApiBoclipsClient } from '../ApiBoclipsClient';
+import { FakeBoclipsClient } from './FakeBoclipsClient';
+
 export * from './FakeBoclipsClient';
 export * from './CollectionsFactory';
 export * from './ContentPartnersFactory';
 export * from './OrderFactory';
+
+export const isATestClient = (
+  client: FakeBoclipsClient | ApiBoclipsClient,
+): client is FakeBoclipsClient => {
+  return (client as FakeBoclipsClient).clear !== undefined;
+};